requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
attrs==23.1.0
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
Flask==3.0.0
Flask-RESTful==0.3.10
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
outcome==1.3.0
protobuf==4.21.12
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.8.0
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
urllib3==2.0.7
Werkzeug==3.0.0
wsproto==1.2.0